requirements.txt
pandas
PyPDF2
python-docx
nltk
scikit-learn
openai
tqdm
beautifulsoup4
requests
spacy
sentence-transformers
email-validator
python-dotenv
pdfplumber
google-api-python-client==2.59.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
